Add SpScanner with Updater that processes results
