+++ rust/src/lib.rs
@@ -4,3 +4,4 @@ mod frb_generated; /* AUTO INJECTED BY flutter_rust_bridge. This line may not be
 mod logger;
 mod stream;
 mod wallet;
+mod scanner;
