+++ rust/src/wallet/mod.rs
@@ -1,3 +1,5 @@
 pub mod recorded;
-pub mod scan;
-pub mod spwallet;
+mod wallet;
+pub mod utils;
+
+pub use wallet::SpWallet;
